Info about this project
Hello! 

This is CRUD project, if you don't know what a CRUD is, I'll tell you: CREATE, READ, UPDATE & DELETE. That's the acronym of 'crud'

But you might be wondering... what's a CRUD?
Well, a crud it's basically a website designed to store data, such as a list of products, a list of name's, files, an e-commerce... whatever u want to store!

What technology is involved in the making of a CRUD? 
For a basic crud: HTML5, CSS, PHP, MySQL and XAMPP (you can use other web servers / alternatives to xampp). Then you'll need to host your website, there are a couple of free hosting websites, in my case, I use 000webhost.

I'm not an expert, so this project might have errors. Nonetheless, this makes me proud!